implement soft_pmap in terms of xmap
